Assign unique IDs to domains.
This fixes a couple of race conditions that arise from the fact
that internally, domains are reused. Now, Domain.Sync.notify
will not notify the wrong domain if the target domain terminates
and its structure is reused.

Also, Domain.join will now wait until the target domain has
fully terminated, so that a subsequent spawn is guaranteed to be
able to reuse the structure.

The new test parallel/domain_id tests these properties.