Track modes in Lambda.lfunction and onwards (#33)
Extends Lambda.function_kind to track modes of curried functions.
The supported modes have a heap-returning prefix, followed by a
local-returning suffix, both of which may be empty. Functions which
do not fit this pattern cannot be expressed as a single Lfunction,
and must be explicitly curried in Lambda and Clambda. (See the new
Lambda.check_lfunction for the exact invariants)

Most of the changes are straightforward plumbing of the new info.
Some delicate changes are needed to preserve the new invariants, in:

  - simplif.ml: optimisations which combine Lfunctions
  - translcore.ml: transl_curried_function and friends
  - closure.ml: optimisations for partial and over-application
  - cmm_helpers.ml: mode-aware variants of caml_curry

Caveat: build_apply in translcore is wrong